server/requirements.txt
cffi==1.15.0
click==8.0.3
cryptography==35.0.0
Flask==2.0.2
Flask-Cors==3.0.10
flask-talisman==0.8.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.1
six==1.16.0
Werkzeug==2.0.2
pymongo[srv]==3.10.1
pydantic>=2.9.2
pydantic_core
pydantic[email]
